Change custom to pre- and post-gain light levels
This allows fine-tuning of the light levels both before and after the preset adjustment curve.

For example, brightening with the pre-gain adjustment and applying the "Crush" mode will increase the lighting contrast of the level.

Note that pairing the pre- and post-gain adjustments with the "Subtract" mode is utterly pointless.